+++ changelog.md
@@ -1,6 +1,16 @@
 # Changelog
 
+## Version 1.4.0 (8 May 2021)
+
+### Added New TreeView Control 
+TODO: Write documentation and an example
+
+### Added New example with_json_cpp_be
+* Use multiple grids on tabbed window
+* GetData from cpp backend class
+
 ## Version 1.3.0 (7 January 2021)
+
 ### Changed column resize and ordering, more mobile friendly
 Now clicking on a column title, an indicator will appear to resize and reorder the column. By clicking again, the indicator will disappear
 
@@ -11,6 +21,7 @@ No more `import BppTableModel 0.1` needed
 [Example](https://dev.bigno.it/bppgrid/color_grid.gif)
 
 ## Version 1.2.0 (11 November 2020)
+
 ### Changed row selection indicator
 Now, all selected row have a background color controlled by properties:
 * `property color dataBkSelOdd`
@@ -32,6 +43,7 @@ If multiselection is enabled `withMultiselection=true` then a new bar under colu
 *WARNING* Deprecated function `getCellBackground(row)` now use: `getCellBk(row, highlight)`
 
 ## Version 1.1.0 (9 April 2020)
+
 ### Add option button on grid
 This button can be disabled see `showOptionsButton`